+++ src/resourcesBase/resourcesBase_domain/src/main/java/com/easondongh/domain/Photo.java
@@ -6,11 +6,12 @@ import java.io.Serializable;
 
 /**
  * 管理图片信息
+ * @author EasonDongH
  */
 @Data
 public class Photo implements Serializable {
 
-    private int id;
+    private long id;
     /**
      * 图片路径
      */
